pyproject.toml
[project]
name = "docai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "abraxas>=0.0.1.dev104",
    "abraxas-predefined>=0.0.1.dev12",
    "pymongo>=4.15.1",
    "strands-agents>=1.8.0",
    "strands-agents-tools>=0.2.0",
]
